// apps/web/src/routes/_dashboard/$user/repos/$repo/-components/__generated__/Commits_history.graphql.ts
export type Commits_history$data = {
  readonly history: {
    readonly edges:
      | ReadonlyArray<
          | {
              readonly node:
                | {
                    readonly abbreviatedOid: string;
                    readonly author:
                      | {
                          readonly email: string | null | undefined;
                          readonly name: string | null | undefined;
                        }
                      | null
                      | undefined;
                    readonly authoredDate: string;
                    readonly committedDate: string;
                    readonly message: string;
                    readonly oid: string;
                    readonly url: string;
                  }
                | null
                | undefined;
            }
          | null
          | undefined
        >
      | null
      | undefined;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null | undefined;
    };
    readonly totalCount: number;
  };
  readonly id: string;
  readonly " $fragmentType": "Commits_history";
};
export type Commits_history$key = {
  readonly " $data"?: Commits_history$data;
  readonly " $fragmentSpreads": FragmentRefs<"Commits_history">;
};